# backend/go/internal/interviewservice/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_interviewservice_usecase_auth_login.Request:
    properties:
      email:
        example: john.smith@gmail.com
        type: string
      password:
        example: not-so-secure-password
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal_interviewservice_usecase_auth_login.Response:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  internal_interviewservice_usecase_auth_register.Request:
    properties:
      confirm_password:
        example: password123
        type: string
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 8
        type: string
    required:
    - confirm_password
    - email
    - password
    type: object
  internal_interviewservice_usecase_servers_getservers.Response:
    properties:
      servers:
        items:
          $ref: '#/definitions/internal_interviewservice_usecase_servers_getservers.ServerItem'
        type: array
    type: object
  internal_interviewservice_usecase_servers_getservers.ServerItem:
    properties:
      id:
        type: string
      infrastructure_resource_id:
        type: string
      power_status:
        type: string
      sku:
        type: string
    type: object
  internal_interviewservice_usecase_servers_power.Request:
    properties:
      action:
        enum:
        - "on"
        - "off"
        type: string
    required:
    - action
    type: object
  internal_interviewservice_usecase_servers_power.Response:
    properties:
      state:
        type: string
      success:
        type: boolean
    type: object
  internal_interviewservice_usecase_servers_provision.Request:
    properties:
      sku:
        type: string
    required:
    - sku
    type: object
  internal_interviewservice_usecase_servers_provision.Response:
    properties:
      id:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 'API สำหรับ Interview Service (Auth: register, login)'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Interview Service API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: เข้าสู่ระบบด้วย email และ password ได้รับ access_token และ refresh_token
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interviewservice_usecase_auth_login.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK - returns access_token and refresh_token in data
          schema:
            $ref: '#/definitions/internal_interviewservice_usecase_auth_login.Response'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized - wrong password
        "404":
          description: Not Found - user not found
        "500":
          description: Internal Server Error
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: สร้างบัญชีผู้ใช้ใหม่
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interviewservice_usecase_auth_register.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request - validation error
        "409":
          description: Conflict - email already exists
        "500":
          description: Internal Server Error
      summary: Register new user
      tags:
      - auth
  /servers:
    get:
      consumes:
      - application/json
      description: รายการ servers ทั้งหมดของ user ที่ login (ต้องส่ง JWT ใน cookie
        หรือ header)
      produces:
      - application/json
      responses:
        "200":
          description: OK - returns servers in data
          schema:
            $ref: '#/definitions/internal_interviewservice_usecase_servers_getservers.Response'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: List servers
      tags:
      - servers
    post:
      consumes:
      - application/json
      description: สร้าง server ใหม่ ต้อง validate SKU กับ infra service (ต้องส่ง
        JWT ใน cookie หรือ header)
      parameters:
      - description: Provision request (sku)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interviewservice_usecase_servers_provision.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK - success and id is Infrastructure Resource ID
          schema:
            $ref: '#/definitions/internal_interviewservice_usecase_servers_provision.Response'
        "400":
          description: Bad Request - invalid or unsupported sku
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "502":
          description: Bad Gateway - infra service error
      security:
      - BearerAuth: []
      summary: Provision server
      tags:
      - servers
  /servers/{server-id}/power:
    post:
      consumes:
      - application/json
      description: เปิดหรือปิด server ต้องส่ง JWT ใน cookie หรือ header
      parameters:
      - description: Server ID (API server id)
        in: path
        name: server-id
        required: true
        type: string
      - description: 'Power action: on or off'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interviewservice_usecase_servers_power.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK - success and current state
          schema:
            $ref: '#/definitions/internal_interviewservice_usecase_servers_power.Response'
        "400":
          description: Bad Request - invalid body or server-id
        "401":
          description: Unauthorized
        "404":
          description: Not Found - server not found
        "500":
          description: Internal Server Error
        "502":
          description: Bad Gateway - infra service error
      security:
      - BearerAuth: []
      summary: Power on/off server
      tags:
      - servers
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"